getwavespeed to return double

+++ trefftz/trefftzwavefe.hpp
@@ -24,7 +24,7 @@ namespace ngfem
                    double aelsize = 1, ELEMENT_TYPE aeltype = ET_TRIG,
                    int abasistype = 0);
 
-    float GetWavespeed () const { return c; }
+    double GetWavespeed () const { return c; }
     void SetWavespeed (double wavespeed) { c = wavespeed; }
 
     // TrefftzWaveFE<D> * SetCenter(Vec<D> acenter) {elcenter = acenter; return
